requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
httpx>=0.27
openai>=1.40
azure-identity>=1.16
pydantic-settings>=2.3

